requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
apscheduler==3.10.4
pydantic==2.7.1
playwright==1.40.0
aiofiles==23.2.1
aiohttp==3.9.1